# docs/swagger.yaml
basePath: /api
definitions:
...
  domain.CreateTransferItemRequest:
    properties:
      product_id:
        example: 550e8400-e29b-41d4-a716-446655440004
        type: string
      qty:
        example: 10
        minimum: 1
        type: integer
    required:
    - product_id
    - qty
    type: object
  domain.CreateTransferRequest:
    properties:
      from_warehouse_id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      items:
        items:
          $ref: '#/definitions/domain.CreateTransferItemRequest'
        type: array
      to_warehouse_id:
        example: 550e8400-e29b-41d4-a716-446655440002
        type: string
    required:
    - from_warehouse_id
    - items
    - to_warehouse_id
    type: object
  domain.TransferStatus:
    enum:
    - REQUESTED
    - APPROVED
    - IN_TRANSIT
    - COMPLETED
    - CANCELLED
    type: string
    x-enum-varnames:
    - TransferStatusRequested
    - TransferStatusApproved
    - TransferStatusInTransit
    - TransferStatusCompleted
    - TransferStatusCancelled
  domain.UpdateTransferStatusRequest:
    properties:
      status:
        allOf:
        - $ref: '#/definitions/domain.TransferStatus'
        example: APPROVED
    required:
    - status
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a comprehensive warehouse management system API for handling
    products, orders, warehouses, and transfers.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Warehouse Management API
  version: "1.0"
paths:
...
  /transfers:
    post:
      consumes:
      - application/json
      description: Create a transfer request to move products between warehouses with
        specified items and quantities
      parameters:
      - description: Transfer creation data with items
        in: body
        name: transfer
        required: true
        schema:
          $ref: '#/definitions/domain.CreateTransferRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Transfer created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request payload or validation failed
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new warehouse transfer
      tags:
      - Warehouse Transfers
  /transfers/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a specific warehouse transfer with all its details and
        items
      parameters:
      - description: Transfer ID (UUID)
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Transfer retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid transfer ID format
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Transfer not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get warehouse transfer by ID
      tags:
      - Warehouse Transfers
  /transfers/{id}/execute:
    post:
      consumes:
      - application/json
      description: Execute an approved transfer to actually move stock between warehouses
      parameters:
      - description: Transfer ID (UUID)
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Transfer executed successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid transfer ID or transfer not in executable state
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Transfer not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error or execution failed
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Execute warehouse transfer
      tags:
      - Warehouse Transfers
  /transfers/{id}/status:
    patch:
      consumes:
      - application/json
      description: Update the status of a warehouse transfer (REQUESTED, APPROVED,
        IN_TRANSIT, COMPLETED, CANCELLED)
      parameters:
      - description: Transfer ID (UUID)
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: New transfer status
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateTransferStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Transfer status updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid transfer ID or status
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Transfer not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update transfer status
      tags:
      - Warehouse Transfers
...
  /warehouses/{warehouse_id}/transfers:
    get:
      consumes:
      - application/json
      description: Retrieve all transfers where the warehouse is either source or
        destination with pagination
      parameters:
      - description: Warehouse ID (UUID)
        format: uuid
        in: path
        name: warehouse_id
        required: true
        type: string
      - default: 10
        description: Number of items per page
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Transfers retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid warehouse ID or pagination parameters
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get transfers by warehouse
      tags:
      - Warehouse Transfers
produces:
- application/json
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"